refactor: move git-ops types into existing vcs/types.ts
Delete redundant git-ops-types.ts — the types belong in the existing
vcs/types.ts file alongside other VCS interfaces.
